fix: `InputMessage.dataLength` to `u32`

+++ packages/transactions/src/coders/input.ts
@@ -284,7 +284,7 @@ export class InputMessageCoder extends Coder<InputMessage, InputMessage> {
     parts.push(new ByteArrayCoder(32).encode(value.nonce));
     parts.push(new NumberCoder('u8').encode(value.witnessIndex));
     parts.push(new U64Coder().encode(value.predicateGasUsed));
-    parts.push(new NumberCoder('u16').encode(data.length));
+    parts.push(new NumberCoder('u32').encode(data.length));
     parts.push(new NumberCoder('u32').encode(value.predicateLength));
     parts.push(new NumberCoder('u32').encode(value.predicateDataLength));
     parts.push(new ByteArrayCoder(data.length).encode(data));
@@ -320,7 +320,7 @@ export class InputMessageCoder extends Coder<InputMessage, InputMessage> {
     const predicateGasUsed = decoded;
     [decoded, o] = new NumberCoder('u32').decode(data, o);
     const predicateLength = decoded;
-    [decoded, o] = new NumberCoder('u16').decode(data, o);
+    [decoded, o] = new NumberCoder('u32').decode(data, o);
     const dataLength = decoded;
     [decoded, o] = new NumberCoder('u32').decode(data, o);
     const predicateDataLength = decoded;
